feat: support multiple type of username deducers
This PR adds support for different types of "username deducers". In its
current version, SSBucketeer simply tries to strip a "user-ssb-" prefix
from the namespace and assumes this is the correct username. This PR
adds a new deducer which can look for an annotation on the namespace.

These can be configured via an environment variables using a yaml
string:

```
SSBUCKETEER_USERNAME_DEDUCERS: |
  - type: annotation
    config:
      annotation: dapla.ssb.no/username
  - type: prefix
    config:
      prefix: "user-ssb-"
```

Using this config, SSBucketeer will first try to look for the
annotation, but fall back to the prefix deducer if it fails.

Closes: #35